Finalyear-Project-Back-end/ai-service/requirements.txt
flask==3.1.1
flask-cors==5.0.1
tensorflow==2.19.0
numpy==2.2.5
pillow==11.3.0
gunicorn==23.0.0
